[lldb-dap] Refactor remaining request handlers (NFC)
