+++ lldb/tools/lldb-dap/Handler/RequestHandler.h
@@ -220,6 +220,76 @@ public:
   void operator()(const llvm::json::Object &request) override;
 };
 
+class PauseRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "pause"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class ScopesRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "scopes"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class SetVariableRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "setVariable"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class SourceRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "source"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class StackTraceRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "stackTrace"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class ThreadsRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "threads"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class VariablesRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "variables"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class LocationsRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "locations"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class DisassembleRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "disassemble"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
+class ReadMemoryRequestHandler : public RequestHandler {
+public:
+  using RequestHandler::RequestHandler;
+  static llvm::StringLiteral getCommand() { return "readMemory"; }
+  void operator()(const llvm::json::Object &request) override;
+};
+
 /// A request used in testing to get the details on all breakpoints that are
 /// currently set in the target. This helps us to test "setBreakpoints" and
 /// "setFunctionBreakpoints" requests to verify we have the correct set of
